Integrate history with cell deletion
