+++ client/components/Editors/Table/index.tsx
@@ -22,7 +22,7 @@ export interface TableEditorProps extends Partial<TypeDataGridProps> {
 }
 
 export default function TableEditor(props: TableEditorProps) {
-  const { source, schema, report, history, selection, gridRef, ...others } = props
+  const { source, schema, report, history, selection, ...others } = props
   const columns = React.useMemo(
     () => createColumns(schema, report, history, selection),
     [schema, report, history, selection]
@@ -30,8 +30,6 @@ export default function TableEditor(props: TableEditorProps) {
   const [rowsPerPage, setRowsPerPage] = React.useState(20)
 
   const [rowHeight] = React.useState(40)
-
-  const [cellSelection, setCellSelection] = React.useState({})
 
   React.useEffect(() => {
     const debouncedHandleResize = debounce(function handleResize() {
@@ -55,23 +53,6 @@ export default function TableEditor(props: TableEditorProps) {
     setRowsPerPage(Math.floor(tableHeight / rowHeight) - 1)
   }
 
-  const deleteCells = async (cells: object) => {
-    for (const [key] of Object.entries(cells)) {
-      const row = key.substring(0, key.indexOf(','))
-      const column = key.substring(key.indexOf(',') + 1, key.length)
-      // the row counting for the method setItemAt starts at 0 and doesn't take into consideration
-      // the header row, this is why we need to substract 2 from the column here
-      await gridRef?.current?.setItemAt(parseInt(row) - 2, { [column]: '' })
-    }
-  }
-
-  const onKeyDown = React.useCallback(
-    (event: { key: any }) => {
-      if (event.key === 'Delete') deleteCells(cellSelection)
-    },
-    [cellSelection, setCellSelection]
-  )
-
   return (
     <InovuaDatagrid
       onReady={resizeTable}
@@ -86,9 +67,6 @@ export default function TableEditor(props: TableEditorProps) {
       limit={rowsPerPage}
       onLimitChange={setRowsPerPage}
       rowHeight={rowHeight}
-      onKeyDown={onKeyDown}
-      defaultCellSelection={cellSelection}
-      onCellSelectionChange={setCellSelection}
       {...others}
     />
   )
